+++ CHANGELOG.rst
@@ -1,6 +1,15 @@
 History
 =======
 
+0.0.4
+-----
+* Add some documentation.
+* Rename `get_marginal_at` to `get_counts_at`. As the function returns raw
+  counts, not a marginal distribution, this is a less misleading name.
+* Don't return the unused `curr_branches_tot` value from `get_edge_weights_at`.
+* Remove `select_next_edge_at`, we need not concern ourselves with problem
+  specific end-user behaviour. We just provide an API to the pseudo-graph.
+
 0.0.3
 -----
 * Add `observations` property for those who may find `crumbs` confusing or odd.